% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.pubchem.bio.R
\name{browse.pubchem.bio}
\alias{browse.pubchem.bio}
\title{browse.pubchem.bio}
\usage{
browse.pubchem.bio(file.name = NULL)
}
\arguments{
\item{file.name}{character filename and full path (i.e. 'C:/my.data/pc.bio.Rdata') of the pubchem.bio dataset you wish to load.  can be either a full pubchem.bio dataset or a taxonomically scored version.}
}
\value{
nothing
}
\description{
an Rshiny dashboard for browsing a pubchem.bio library. enables filtering by mass/adduct, name, viewing structures, and calculating expected isotope envelopes.
}
\details{
just a viewer, so returns nothing, but enables interactivity with the data to support manual annotation efforts.
}
\author{
Corey Broeckling
}
