% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.cid.lca.R
\name{build.cid.lca}
\alias{build.cid.lca}
\title{build.cid.lca}
\usage{
build.cid.lca(
  pc.directory = NULL,
  tax.sources = "LOTUS - the natural products occurrence database",
  use.pathways = TRUE,
  use.conserved.pathways = FALSE,
  threads = 8,
  verbose = FALSE
)
}
\arguments{
\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{tax.sources}{vector. which taxonomy sources should be used?  defaults to c("LOTUS - the natural products occurrence database", "The Natural Products Atlas", "KNApSAcK Species-Metabolite Database", "Natural Product Activity and Species Source (NPASS)").}

\item{use.pathways}{logical.  default = TRUE, should pathway data be used in building lowest common ancestor, when taxonomy is associated with a pathway?}

\item{use.conserved.pathways}{logical. default = FALSE, should 'conserved' pathways be used?  when false, only pathways with an assigned taxonomy are used.}

\item{threads}{integer.  number of threads to use when finding lowest common ancestor.  parallel processing via DoParallel and foreach packages.}

\item{verbose}{logical. if TRUE, print summary progress to console.}
}
\value{
a data frame containing pubchem CID ('cid'), and lowest common ancestor ('lca') NCBI taxonomy ID integer. will also save to pc.directory as .Rdata file.
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' as input to generate a relationship between pubchem CID and the lowest common ancestor NCBI taxid
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
