% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.pubchem.bio.R
\name{build.pubchem.bio}
\alias{build.pubchem.bio}
\title{build.pubchem.bio}
\usage{
build.pubchem.bio(
  pc.directory = NULL,
  use.bio.sources = TRUE,
  bio.sources = c("Metabolomics Workbench", "Human Metabolome Database (HMDB)", "ChEBI",
    "LIPID MAPS", "MassBank of North America (MoNA)"),
  use.pathways = TRUE,
  pathway.sources = NULL,
  use.taxid = TRUE,
  taxonomy.sources = NULL,
  use.parent.cid = TRUE,
  remove.salts = TRUE,
  get.properties = TRUE,
  threads = 8,
  rcdk.desc = c("org.openscience.cdk.qsar.descriptors.molecular.XLogPDescriptor",
    "org.openscience.cdk.qsar.descriptors.molecular.AcidicGroupCountDescriptor",
    "org.openscience.cdk.qsar.descriptors.molecular.BasicGroupCountDescriptor",
    "org.openscience.cdk.qsar.descriptors.molecular.TPSADescriptor")
)
}
\arguments{
\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{use.bio.sources}{logical.  If TRUE (default) use the bio.source vector of sources, incorporating all CIDs from those bio databases.}

\item{bio.sources}{vector of source names from which to extract pubchem CIDs.  all can be found here: https://pubchem.ncbi.nlm.nih.gov/sources/.  deafults to c("Metabolomics Workbench", "Human Metabolome Database (HMDB)", "ChEBI", "LIPID MAPS",  "MassBank of North America (MoNA)")}

\item{use.pathways}{logical.  should all CIDs from any biological pathway data be incorporated into database?}

\item{pathway.sources}{character. vector of sources to be used when adding metabolites to pubchem bio database. default = NULL, using all pathway sources.}

\item{use.taxid}{logical.  should all CIDs associated with a taxonomic identifier (taxid) be used?}

\item{taxonomy.sources}{character. vector of sources to be used when adding taxonomically related metabolites to database.  Default = NULL, using all sources.}

\item{use.parent.cid}{logical. should CIDs be replaced with parent CIDs?  default = TRUE.}

\item{remove.salts}{logical.  should salts be removed from dataset?  default = TRUE.  salts recognized as '.' in smiles string.  performed after 'use.parent.cid'.}

\item{get.properties}{logical. if TRUE, will return rcdk calculated properties:  XLogP, TPSA, HBondDonorCount and HBondAcceptorCount.}

\item{threads}{integer. how many threads to use when calculating rcdk properties.  parallel processing via DoParallel and foreach packages.}

\item{rcdk.desc}{vector. character vector of valid rcdk descriptors.  default = rcdk.desc <- c("org.openscience.cdk.qsar.descriptors.molecular.XLogPDescriptor", "org.openscience.cdk.qsar.descriptors.molecular.AcidicGroupCountDescriptor", "org.openscience.cdk.qsar.descriptors.molecular.BasicGroupCountDescriptor", "org.openscience.cdk.qsar.descriptors.molecular.TPSADescriptor"). To see descriptor categories: 'dc <- rcdk::get.desc.categories(); dc' .  To see the descriptors within one category: 'dn <- rcdk::get.desc.names(dc[4]); dn'. Note that the four default parameters are relatively fast to calculate - some descriptors take a very long time to calculate.  you can calculate as many as you wish, but processing time will increase the more descriptors are added.}
}
\value{
a data frame containing pubchem CID, title, formula, monoisotopic molecular weight, inchikey, smiles, cas, optionally pubchem properties
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
