% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.taxid.lca.R
\name{build.taxid.lca}
\alias{build.taxid.lca}
\title{tax.filter.pubchem.bio}
\usage{
build.taxid.lca(
  pc.directory = NULL,
  tax.sources = NULL,
  use.pathways = TRUE,
  use.parent.cid = TRUE,
  threads = 8
)
}
\arguments{
\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{tax.sources}{character vector.  containing taxonmy source values (from NCBI taxonomy database). i.e. c("LOTUS - the natural products occurrence database", "The Natural Products Atlas").  if NULL (default), use all values.}

\item{use.pathways}{character vector.  default = NULL, all sources used.  can set to 'interactive' to enable you to choose the sources at the console.  or set to source name vector, i.e. c("LOTUS - the natural products occurrence database", "The Natural Products Atlas").}

\item{use.parent.cid}{logical. should CIDs be replaced with parent CIDs?}

\item{threads}{integer. how many threads to use when calculating lowest common ancestor.  parallel processing via DoParallel and foreach packages.}
}
\value{
a data frame containing pubchem CID ('cid'), and lowest common ancestor ('lca') NCBI taxonomy ID integer. will also save to pc.directory as .Rdata file.
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function to filter a dataset created by 'build.pubchem.bio' function
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
