% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.accurate.mass}
\alias{cid.accurate.mass}
\title{cid.accurate.mass.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.accurate.mass file from get.pubchem.ftp
}
\description{
A subset of the full cid.accurate.mass, for example code
}
