% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.cas}
\alias{cid.cas}
\title{cid.cas.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.cas file from get.pubchem.ftp
}
\description{
A subset of the full cid.cas, for example code
}
