% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.formula}
\alias{cid.formula}
\title{cid.formula.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.formula file from get.pubchem.ftp
}
\description{
A subset of the full cid.formula, for example code
}
