% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.inchi}
\alias{cid.inchi}
\title{cid.inchi.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.inchi file from get.pubchem.ftp
}
\description{
A subset of the full cid.inchi, for example code
}
