% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.inchikey}
\alias{cid.inchikey}
\title{cid.inchikey.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.inchikey file from get.pubchem.ftp
}
\description{
A subset of the full cid.inchikey, for example code
}
