% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.lca}
\alias{cid.lca}
\title{cid.lca.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.lca file from get.pubchem.ftp
}
\description{
A subset of the full cid.lca, for example code
}
