% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.mesh.function}
\alias{cid.mesh.function}
\title{cid.mesh.function.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.mesh file from get.pubchem.ftp
}
\description{
A subset of the full cid.mesh, for example code
}
