% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.mesh.name}
\alias{cid.mesh.name}
\title{cid.mesh.name.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.mesh.name file from get.pubchem.ftp
}
\description{
A subset of the full cid.mesh.name, for example code
}
