% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.monoisotopic}
\alias{cid.monoisotopic}
\title{cid.monoisotopic.mass.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.monoisotopic file from get.pubchem.ftp
}
\description{
A subset of the full cid.monoisotopic, for example code
}
