% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.parent}
\alias{cid.parent}
\title{cid.parent.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.parent file from get.pubchem.ftp
}
\description{
A subset of the full cid.parent, for example code
}
