% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.pmid.ct}
\alias{cid.pmid.ct}
\title{cid.pmid.ct.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.pmid.ct file from get.pubchem.ftp
}
\description{
A subset of the full cid.pmid.ct, for example code
}
