% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.pmid}
\alias{cid.pmid}
\title{cid.pmid.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.pmid file from get.pubchem.ftp
}
\description{
A subset of the full cid.pmid, for example code
}
