% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.preferred}
\alias{cid.preferred}
\title{cid.preferred.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.preferred file from get.pubchem.ftp
}
\description{
A subset of the full cid.preferred, for example code
}
