% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.pwid}
\alias{cid.pwid}
\title{cid.pwid.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.pwid file from get.pubchem.ftp
}
\description{
A subset of the full cid.pwid, for example code
}
