% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.sid}
\alias{cid.sid}
\title{cid.sid.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.sid file from get.pubchem.ftp
}
\description{
A subset of the full cid.sid, for example code
}
