% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.smiles}
\alias{cid.smiles}
\title{cid.smiles.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.smiles file from get.pubchem.ftp
}
\description{
A subset of the full cid.smiles, for example code
}
