% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.synonym}
\alias{cid.synonym}
\title{cid.synonym.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.synonym file from get.pubchem.ftp
}
\description{
A subset of the full cid.synonym, for example code
}
