% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.taxid}
\alias{cid.taxid}
\title{cid.taxid.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.taxid file from get.pubchem.ftp
}
\description{
A subset of the full cid.taxid, for example code
}
