% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cid.title}
\alias{cid.title}
\title{cid.title.rda}
\format{
data.table, stored in .rda format
}
\source{
subset of cid.title file from get.pubchem.ftp
}
\description{
A subset of the full cid.title, for example code
}
