% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.pubchem.bio.R
\name{export.pubchem.bio}
\alias{export.pubchem.bio}
\title{export.pubchem.bio}
\usage{
export.pubchem.bio(pc.bio.object = NULL, export.file.name = NULL)
}
\arguments{
\item{pc.bio.object}{input data.table, generated from 'build.pubchem.bio' or 'build.taxon.metabolome' functions}

\item{export.file.name}{valid file path and name.  Extension should be listed as '.tsv'.}
}
\value{
nothing - file written to disk.
}
\description{
export pubchem.bio pc.bio syle data.table to tab delimited text file for import into other programs.  all columns exported.
}
\details{
takes output from 'build.pubchem.bio' or 'build.taxon.metabolome' functions, reformatting, and exporting to input format suitable for MSFinder.
}
\author{
Corey Broeckling
}
