% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.taxon.metabolome.R
\name{get.taxon.metabolome}
\alias{get.taxon.metabolome}
\title{get.taxon.metabolome}
\usage{
get.taxon.metabolome(
  pc.directory = NULL,
  taxid = c(),
  get.properties = TRUE,
  threads = 4
)
}
\arguments{
\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{taxid}{integer vector of integer NCBI taxonomy IDs.  i.e.  c(9606, 1425170 ) for Homo sapiens and Homo heidelbergensis.}

\item{get.properties}{logical. if TRUE, will return rcdk calculated properties:  XLogP, TPSA, HBondDonorCount and HBondAcceptorCount.}

\item{threads}{integer. how many threads to use when calculating rcdk properties.  parallel processing via DoParallel and foreach packages.}
}
\value{
a data frame containing pubchem CID ('cid'), and lowest common ancestor ('lca') NCBI taxonomy ID integer. will also save to pc.directory as .Rdata file.
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function to filter a dataset created by 'build.pubchem.bio' function
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
