% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pc.bio.subset}
\alias{pc.bio.subset}
\title{pc.bio.subset.rda}
\format{
data.table, stored in .rda format
}
\source{
pubchem.bio data.table output derived from build.pubchem.bio function
}
\description{
A small dataset of a pubchem.bio metabolome scored by taxon, for inclusion in vignette
}
