% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{.Rda}
\name{pc.bio.tax.scored.subset.Rda}
\alias{pc.bio.tax.scored.subset.Rda}
\title{A small dataset of a pubchem.bio metabolome scored by taxon, for inclusion in vignette}
\description{
A small dataset of a pubchem.bio metabolome scored by taxon, for inclusion in vignette
}
\author{
Corey Broeckling
}
\keyword{data}
\keyword{pubchem.bio}
