% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sub.taxid.hierarchy}
\alias{sub.taxid.hierarchy}
\title{sub.taxid.hierarchy.rda}
\format{
data.table, stored in .rda format
}
\source{
pubchem.bio data.table output from NCBI Taxonomy data
}
\description{
A small dataset of a the taxonomy hierarchy, for inclusion in vignette
}
