% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{.Rda}
\name{sub.taxid.hierarchy.Rda}
\alias{sub.taxid.hierarchy.Rda}
\title{A small dataset of a the taxonomy hierarchy, for inclusion in vignette}
\description{
A small dataset of a the taxonomy hierarchy, for inclusion in vignette
}
\author{
Corey Broeckling
}
\keyword{data}
\keyword{pubchem.bio}
