% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{taxid.hierarchy}
\alias{taxid.hierarchy}
\title{taxid.hierarchy.rda}
\format{
data.table, stored in .rda format
}
\source{
pubchem.bio data.table output from NCBI Taxonomy data
}
\description{
A subset of the full taxid.hierarchy, for example code
}
